feat(cli): F3 收口(#121) — 抽 projection 层 + pickLang 单一来源
issue #121 第 3 组(#7 主体 / #18 partial):

- 新建 src/cli/oclif/projection.ts:projectCommand(cmd, lang) / projectTopic(topic, lang) 两个 helper,把 Command.Loadable → 单语视图的 clone + pickLang 逻辑收口在一处。reviewer P1-A 的关切是「build-docs.ts 跟 help.ts 各做一遍 reflection,oclif 升级新字段静默漂移」— projection 拿出来后,future 加 hidden / aliases / deprecationOptions 只改一处。
- src/cli/oclif/help.ts:删 inline filterCommand / filterTopic,改用 projection。LangAwareHelp 从 90 行缩到 35 行,纯薄壳。
- scripts/build-docs.ts:inline pickLang 改 import src/cli/oclif/i18n.ts 的 pickLang(单一来源),保留 build-docs 自己的 `?? ''` wrapper(build-docs 拼字符串需要 string 不是 string | undefined)。

scope 决策:#18 build-docs.ts 拆 lib + index(460 行重排)+ LangAwareHelp.formatTopics 专项测试 — skip。pickLang 单一来源 + projection 抽出已经解决 reviewer 「reflection 重复 + oclif 升级漂移」的核心关切;build-docs 拆分是结构改善,在 #121 close 范围外可独立追踪。formatTopics lang 切换由 init hook mutate + 现有 e2e(oclif-hook-init.test.ts)间接覆盖,专项 in-process 测试增量价值有限。
